Haszowanie haseł w DB. Trzeba zpullować DB z haszowanymi hasłami, login i hasło te same.

+++ src/main/java/pap/Item.java
@@ -10,7 +10,7 @@ public class Item {
     String description;
 
 
-    public Item(String name, double price){
+    public Item(String name, int price){
         this.name = name;
         this.price = price;
     }
